Extract duplicated code to its own class
Both of these test classes had duplicated code that would need to be updated, so easier to extract to a common place first
